Export `decorateStory` from frameworks that have it.
And no longer emit warnings for exported `decorateStory` or `renderToDOM`
