+++ app/svelte/src/client/preview/config.ts
@@ -1,3 +1,4 @@
 export { renderToDOM } from './render';
+export { decorateStory } from './decorators';
 
 export const parameters = { framework: 'svelte' };
