+++ lib/builder-webpack4/src/preview/virtualModuleEntry.template.js
@@ -41,6 +41,10 @@ Object.keys(config).forEach((key) => {
       v[key] = value;
       return addParameters(v, false);
     }
+    case 'decorateStory':
+    case 'renderToDOM': {
+      return null; // This key is not handled directly in v6 mode.
+    }
     default: {
       // eslint-disable-next-line prefer-template
       return console.log(key + ' was not supported :( !');
